# Simplify bash_completions.go