Use buf instead of writing directly to w.
Improve test coverage from 80% to 85%.