@eparis 2017-05-14
I don't love it. It definitely fixes some dumb things like: `b := (len(flag.NoOptDefVal) > 0; if !b {` (which I probably wrote and feel badly about looking back on it.

But most of this patch is just removing error handling code by storing everything in RAM instead of using the writer we were given. And eventually you have to have the data in RAM twice while you copy it from the buffer to the writer. Why is that an improvement? The testing % went up, but now, in a memory constrained environment we will panic in all the places we would have had an error.

It's not a nak but it is also not a LGTM from me. There are definitely a couple of things which are better in here. I'm neither going to block the merge nor click the button myself.