@bogem 2017-05-14
> Why is that an improvement?

```
$ benchcmp before after
benchmark                     old ns/op     new ns/op     delta
BenchmarkBashCompletion-4     450778        120105        -73.36%

benchmark                     old allocs     new allocs     delta
BenchmarkBashCompletion-4     350            472            +34.86%

benchmark                     old bytes     new bytes     delta
BenchmarkBashCompletion-4     10877         36761         +237.97%
```
Yes, the RAM usage increased, but only from 11KB to 37KB. I don't think it's a great deal for PCs in 2017. But keep in mind, what performance increased in 3.7 times ...from 460μs to 120μs, not so much too.

> in  a memory constrained environment we will panic in all the places we would have had an error
I don't think, that someone's PC will not have ~20KB for this. 

The main thing is not in performance, but in this code style. I hate code, where there are `if _, err := fmt.Fprint(...); err != nil` in almost every line. With my changes code looks much better and more understandable. You don't have to consider these `if _, err := fmt.Fprint(...); err != nil` lines and we can focus more on more important aspects of bash completion generation.