@eparis 2017-05-18
I stick with my 'I don't love it'.  But I understand its appeal. Feel free to merge.